Move decode of transaction to within constructor
